removed useless class in chain parser with only static methods
