+++ mhl/history.py
@@ -13,8 +13,7 @@ import os
 import re
 
 from .__version__ import ascmhl_folder_name, ascmhl_file_extension, ascmhl_chainfile_name
-from .chain_txt_backend import MHLChainTXTBackend
-from . import hashlist_xml_parser
+from . import hashlist_xml_parser, chain_txt_parser
 from .utils import datetime_now_filename_string
 from typing import Tuple, List, Dict, Optional, Set
 from . import logger
@@ -172,7 +171,7 @@ class MHLHistory:
         history.asc_mhl_path = asc_mhl_folder_path
 
         file_path = os.path.join(asc_mhl_folder_path, ascmhl_chainfile_name)
-        history.chain = MHLChainTXTBackend.parse(file_path)
+        history.chain = chain_txt_parser.parse(file_path)
 
         hash_lists = []
         for root, directories, filenames in os.walk(asc_mhl_folder_path):
